+++ petugas/show_masyarakat.php
@@ -2,7 +2,9 @@
 	require '../koneksi.php';
 	if (!isset($_SESSION['id_petugas'])) {
 		header("Location: login_petugas.php");
+		exit();
 	}
+	
 	$masyarakat = mysqli_query($koneksi, "SELECT * FROM masyarakat");
 ?>
 
@@ -15,8 +17,10 @@
 </head>
 <body>
 	<?php include 'sidebar.php'; ?>
-
-	<a href="insert_masyarakat.php">Tambah Masyarakat</a>
+	<?php if ($_SESSION['level'] == 'admin'): ?>
+		<a href="insert_masyarakat.php">Tambah Masyarakat</a>
+	<?php endif ?>	
+	
 	<table border="1" cellpadding="10" cellspacing="0">
 		<thead>
 			<tr>
@@ -26,7 +30,9 @@
 				<th>Username</th>
 				<th>No. Telepon</th>
 				<th>Alamat</th>
-				<th>Aksi</th>
+				<?php if ($_SESSION['level'] == 'admin'): ?>
+					<th>Aksi</th>
+				<?php endif ?>
 			</tr>
 		</thead>
 		<tbody>
@@ -39,10 +45,12 @@
 					<td><?= $dataMasyarakat['username']; ?></td>
 					<td><?= $dataMasyarakat['telp']; ?></td>
 					<td><?= $dataMasyarakat['alamat']; ?></td>
-					<td>
-						<a href="update_masyarakat.php?nik=<?= $dataMasyarakat['nik']; ?>">Ubah</a>
-						<a onclick="return confirm('Apakah anda yakin menghapus data masyarakat dengan NIK <?= $dataMasyarakat['nik']; ?>?')" href="delete_masyarakat.php?nik=<?= $dataMasyarakat['nik']; ?>">Hapus</a>
-					</td>
+					<?php if ($_SESSION['level'] == 'admin'): ?>
+						<td>
+							<a href="update_masyarakat.php?nik=<?= $dataMasyarakat['nik']; ?>">Ubah</a>
+							<a onclick="return confirm('Apakah anda yakin menghapus data masyarakat dengan NIK <?= $dataMasyarakat['nik']; ?>?')" href="delete_masyarakat.php?nik=<?= $dataMasyarakat['nik']; ?>">Hapus</a>
+						</td>
+					<?php endif ?>
 				</tr>
 			<?php endforeach ?>
 		</tbody>
