+++ petugas/show_pengaduan.php
@@ -2,7 +2,9 @@
 	require '../koneksi.php';
 	if (!isset($_SESSION['id_petugas'])) {
 		header("Location: login_petugas.php");
+		exit();
 	}
+
 	$pengaduan = mysqli_query($koneksi, "SELECT * FROM pengaduan INNER JOIN masyarakat ON pengaduan.nik = masyarakat.nik INNER JOIN kelurahan ON pengaduan.id_kelurahan = kelurahan.id_kelurahan");
 ?>
 
@@ -15,8 +17,9 @@
 </head>
 <body>
 	<?php include 'sidebar.php'; ?>
-
-	<a href="insert_pengaduan.php">Tambah Pengaduan</a>
+	<?php if ($_SESSION['level'] == 'admin'): ?>
+		<a href="insert_pengaduan.php">Tambah Pengaduan</a>
+	<?php endif ?>
 	<table border="1" cellpadding="10" cellspacing="0">
 		<thead>
 			<tr>
@@ -42,8 +45,15 @@
 					<td><?= $dataPengaduan['status']; ?></td>
 					<td><?= $dataPengaduan['kelurahan']; ?></td>
 					<td>
-						<a href="update_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Ubah</a>
-						<a onclick="return confirm('Apakah anda yakin menghapus data pengaduan dengan isi laporan <?= $dataPengaduan['isi_laporan']; ?>?')" href="delete_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Hapus</a>
+						<?php if ($dataPengaduan['status'] == 'proses'): ?>
+							<a href="insert_tanggapan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Tanggapi</a>
+						<?php endif ?>
+						
+						<?php if ($_SESSION['level'] == 'admin'): ?>
+							<a href="update_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Ubah</a>
+							<a onclick="return confirm('Apakah anda yakin menghapus data pengaduan dengan isi laporan <?= $dataPengaduan['isi_laporan']; ?>?')" href="delete_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Hapus</a>
+						<?php endif ?>
+
 					</td>
 				</tr>
 			<?php endforeach ?>
